chromebook-setup: add support for a path as storage.
In order to use nfsroot, we want to be able to store the root filesystem
and the the kernel on a non-block device. This patch adds support to do
this using the same parameters. If a block device is detected in
--storage option the rootfs is writen to the second partition of the
device and the kernel dd'ied to the first partition. If a non-block
device is detected the rootfs is writen to the specified path and the
kernel in $ROOTFS_DIR/boot.

Signed-off-by: Ezequiel Garcia <ezequiel@collabora.co.uk>
[Rework the original patch to support a path as storage]
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>