chromebook-setup: Increase kernel partition to 32MB
Usually 16MB is enough for the kernel size and even sometimes on old
bootloaders more than 16MB is not supported, but we can exceed that
limit when we want, for example, enable some debugging functionality.
Increase the kernel partition to 32MB so we can at least run a kernel
with debug information on those devices that support it.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>