bootstub: Add bootstub.efi file to avoid building it
It is known that build bootstub.efi is broken on differents distros like
Fedora. Fixing this for every distro might make no extra value, so,
instead, just add the bootstub.efi file generated with a Debian distro
because that should just work.

This commit does not removes the bootstub source code just in case it
will be helpful in the future and if, for some reason, we need to build
again the bootstub.efi file.

Reported-by: Fabien Lahoudere <fabien.lahoudere@collabora.com>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>