testing: chromebooks-tpl.yaml: Add board profiles and use configurable variables
lqa is able to replace the YAML template variables like {{variable}}
with the values specified using -t options from the command line. Use
those variables allows us to make the YAML file generic for all the
supported boards in the LAVA Lab. Add a profiles file and the variables
needed in order to be able to use the same YAML file to launch a LAVA
job for a specific board or all boards. I.e: You should be able to see the
jobs running issuing a:

  lqa submit -g chromebooks-profiles.yaml --all-profiles -t kernel_version:<kernel version> chromebooks-tpl.yaml

Or instead of using --all-profiles use -p <device_type> to launch a LAVA
job for a specific device.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>