chromebook-setup: Add deploy kernel only command
Sometimes we just want to update the kernel fit image,
e.g. change the device tree or the kernel boot comand.
In this case we don't need to copy again the whole kernel modules
which takes a lot of time. Option deploy_kernel_only just
builds a fit image and writes it to the storage medium.

Signed-off-by: Matthias Brugger <mbrugger@suse.com>