chromebook-setup: Add an extra kernel parameters option
On some devices, like Cherrytrail devices, we will need to pass extra
parameter to the kernel in order to fix some issues. For that purpose
add and extra kernel parameter option to be able to pass these
parameters.

Pixel and others devices has known bug with its TPM where it doesn't
initialize correctly. To get around this, the tpm_tis module must be
force-loaded at boot. If this is not done, the Chromebook performs a
hard reboot when attempting to resume from suspend. While here,
force-loaded at boot for AMD64 based devices.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>