chromebook-setup: Allow to specify the arch exporting an ARCH variable
A common convention for build systems that support cross-compilation is to
use the ARCH and CROSS_COMPILE environment variables to determine the arch
and toolchain to use.

The script already supports the CROSS_COMPILE environment variable, but it
requires to specify the architecture using an --architecture option.

Allow the user to also specify the architecture using the ARCH environment
variable, and only make the --architecture option required if the variable
was not set.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>