fragments: Add a Fedora Kconfig fragment
Some distros have custom configuration that the kernel needs to support.

For example, Fedora ships the firmware files compressed to reduce the disk
size needed so the option to support compressed firmwares must be enabled.

Also, that distro uses zram backed swap so it has to be enabled as module
along with the compression algorithms.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>