chromebook-setup: Add experimental support to create Fedora based images
Add a new deploy_fedora command that is able to create a bootable media
device for some Chromebooks based on Fedora rootfs. Right now the kernel
install is downloaded from a personal copr repository but we plan to use
the image kernel in the future. As example, the following command

  ./chromebook-setup.sh deploy_fedora --architecture=arm64 --storage=/dev/sdX

creates a bootable media image based on Fedora 36 that works on some Qualcomm
Chromebooks.

Signed-off-by: Dorinda Bassey <dbassey@redhat.com>
Signed-off-by: Enric Balletbo i Serra <eballetbo@redhat.com>