chromebook.install: Copy the kernel image to the boot directory
Other kernel-install scripts assume that this would be done by a previous
script (e.g: /usr/lib/kernel/install.d/20-grub.script), but the grub2 pkg
won't be installed in a Chromebook where Coreboot is used as a bootloader.

So copying the kernel image in the boot directory should be handled by the
script as well or otherwise the FIT image generation will fail.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>