chromebook.install: Improve heuristic to choose writting partition
Currently just the first one that has a type kernel is used. But a better
heuristic is to use a partition with type kernel on the same device that
contains the root partition mounted.

Ohterwise for example it will write to kernel partitions in the eMMC even
when the machine was booted using a USB disk.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>