chromebook-setup: Don't hardcode the latest image in the URL
Currently the GETFEDORA variable in the script configuration should be set
to a specific image, but it is more flexible to instead set it to a parent
diretory and let the script determine the latest image for download.

This can be useful for example when downloading images whose filenames are
not static, which is the case for Rawhide images since these are generated
on a daily basis.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>