chromebook-setup: Use the partition with label ROOT-A as rootfs
Use the partition with label ROOT-A as the rootfs partition, since that is
created by the script when the block device is partitioned.

Also, there is a single rootfs partition in the used layout so the rootfs
partition can be a fixed one.

This also fixes a boot issue when using a Fedora initrd to boot, because
dracut doesn't understand the PARTUUID=%U/PARTNROFF=1 idiom used before:

[ TIME ] Timed out waiting for device 60-8035c990b362\x2fPARTNROFF=1.
[DEPEND] Dependency failed for init…e.target - Initrd Root Device.
[DEPEND] Dependency failed for syst…-8e60-8035c990b362\x2fPARTNROFF=1.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>