chromebook-setup: Use the rootfs UUID as root param instead of a label
Instead of using the ROOT-A label to find the partition to mount as rootfs
use the filesystem UUID if the root partition is known. Use the label only
as a fallback if the root partition is not known (i.e: for build_vboot and
others command that only deal with the vboot).

This is more reliable because otherwise if there are more than one device
with a partition using a ROOT-A label, the wrong partition could be found.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>