chromebook-setup: Set a default kernel path in cmd_get_kernel()
Instead of doing that at the start of the script as a part of the options
sanitising because it may be important for other commands to know whether
a --kernel option was passed to the script.

A default kernel path is only needed for the cmd_get_kernel() function to
have a default path to clone the Linux kernel repository. So only set the
default path if that function is called.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>