chromebook-setup: Decouple --kernel from CB_KERNEL_PATH variable
The CB_KERNEL_PATH variable is set to the argument of the --kernel option,
or to a default "kernel" path if that option was not used. But there are
cases where the script would like to know whether the --kernel option was
used or not.

Since the same variable is used to denote both things, it is not possible
for the script to determine that. Decouple the value of --kernel argument
with the kernel path, to avoid the former value to be lost.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>