chromebook-setup: Don't assume that always an initramfs is present
If we try to build a FIT image using a Fedora kernel, it can be assumed that an
initramfs will always be present in arch/${ARCH}/boot/initramfs-$kernel_version.img,
otherwise, is possible that an initramfs won't be used, i.e. when you call
cmd_build_kernel() from a custom kernel build.

This patch fixes the previous assumption that always an initramfs is
present and makes the commands like cmd_build_kernel() to work again.

Signed-off-by: Enric Balletbo i Serra <eballetbo@redhat.com>