chromebook-setup: Use PARTUUID to pass the root partition
When using UUIDs, PARTUUID, which is stored in the partition table IS the only way that is
available at boot time. `root=LABEL=` and `root=UUID=` only works with an initramfs that
fetches these identifiers. Ideally we should have here `root=PARTUUID=%U/PARTNROFF=1` where
%U is passed by the bootloader, but, as dracut doesn't support the PARTNROFF we stick on the
PARTUUID of the rootfs partition.

Signed-off-by: Enric Balletbo i Serra <eballetbo@redhat.com>