chromebook-setup: Remove the root password
If we want to login in the console, apart from unlock the root account
we need to add or remove a root password. As this is just a development
tool and experimental steps to support Fedora images, at the moment,
leave the root password empty so an user can login via serial connection
when the display doesn't work.

Signed-off-by: Enric Balletbo i Serra <eballetbo@redhat.com>