chromebook-setup: Move the temp dir creation to a helper function
If there is an error and the temporary directory is not cleaned up, the
script fails with:

  mkdir: cannot create directory ‘./tmpdir’: File exists

Commit 1bbf2e1708f8 ("chromebook-setup: mkdir fails if directory already
exists") fixed one instance of this, but cmd_deploy_fedora_kernel() also
crates a tmp dir and had the same issue.

Let's add a helper function to create the temp dir in a single place.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>