chromebook-setup: Switch to use a Fedora rootfs by default
The Debian images are not available anymore and its being a while since
we are using a Fedora rootfs. Using the official raw images provided by
Fedora is more scalable than using the custom Debian ones, so switch to
use a Fedora rootfs by default and remove the unavailable Debian image
references.

Signed-off-by: Enric Balletbo i Serra <eballetbo@redhat.com>