chromebook-setup: Use proper qemu-aarch64-static path
The chroot tree must use the same path as is registered in binfmt-misc.  Use
that path rather than hard-wired /usr/bin/qemu-aarch64-static.

This also ensures that the needed binfmt-misc registration exists.  If it
doesn't, it can cause a lot of confusion to an unaware user.

Note that some precaution is needed to not swap aarch64 with aarch64_be (or
another aarch64*) if present; hence using wildcards for simplification is not a
good idea.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>